design improved, need to fix some minor problems
