+++ src/main/webapp/WEB-INF/pages/login.jsp
@@ -8,23 +8,110 @@
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
 <!DOCTYPE html>
 <html lang="fr">
+
 <head>
-    <title>Page de connexion</title>
+
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="description" content="">
+    <meta name="author" content="">
+
+    <title>Projet01 - AMT 2016-2017</title>
+
+    <!-- Bootstrap Core CSS -->
+    <link href="css/bootstrap.min.css" rel="stylesheet">
+
+    <!-- Custom CSS -->
+    <link href="css/landing-page.css" rel="stylesheet">
+
+    <!-- Custom Fonts -->
+    <link href="font-awesome/css/font-awesome.min.css" rel="stylesheet" type="text/css">
+    <link href="https://fonts.googleapis.com/css?family=Lato:300,400,700,300italic,400italic,700italic" rel="stylesheet" type="text/css">
+
+    <!-- HTML5 Shim and Respond.js IE8 support of HTML5 elements and media queries -->
+    <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
+    <!--[if lt IE 9]>
+    <script src="https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>
+    <script src="https://oss.maxcdn.com/libs/respond.js/1.4.2/respond.min.js"></script>
+    <![endif]-->
+
 </head>
+
 <body>
-<h1>Connexion</h1>
-<p>Veuillez-vous connecter.</p>
-<form method="post" action="">
-    <p>
-        <label for="userName">Votre nom d'utilisateur (email) : </label><input type="text" name="userName" id="userName" />
-        <br>
-        <label for="userPassword">Votre mot de passe : </label><input type="password" name="userPassword" id="userPassword" />
-        <br>
-        <input type="submit" value="Envoyer" />
-    </p>
-</form>
-<p>
-    <a href="register">Créer un compte</a>
-</p>
+
+<!-- Navigation -->
+<nav class="navbar navbar-default navbar-fixed-top topnav" role="navigation">
+    <div class="container topnav">
+        <!-- Brand and toggle get grouped for better mobile display -->
+        <div class="navbar-header">
+            <button type="button" class="navbar-toggle" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1">
+                <span class="sr-only">Navigation</span>
+                <span class="icon-bar"></span>
+                <span class="icon-bar"></span>
+                <span class="icon-bar"></span>
+            </button>
+        </div>
+        <!-- Collect the nav links, forms, and other content for toggling -->
+        <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
+            <ul class="nav navbar-nav navbar-right">
+                <li>
+                    <a href="/Projet01">Retourner à la page d'accueil</a>
+                </li>
+                <!-- <li>
+                    <a href="#services">Services</a>
+                </li>
+                <li>
+                    <a href="#contact">Contact</a>
+                </li> -->
+            </ul>
+        </div>
+        <!-- /.navbar-collapse -->
+    </div>
+    <!-- /.container -->
+</nav>
+
+<!-- Page Content -->
+
+<div class="content-section-b">
+
+    <div class="container">
+
+        <div class="row">
+            <div class="col-lg-5 col-lg-offset-1 col-sm-push-6  col-sm-6">
+                <div class="clearfix"></div>
+                <h2 class="section-heading">Connexion</h2>
+                <p class="lead">Veuillez fournir vos informations de connexion.</p>
+                <form id="loginForm" method="post" action="#">
+                    <div class="form-group">
+                        <label for="userName">Nom d'utilisateur (email) :</label>
+                        <input type="email" class="form-control" id="userName" name="userName" placeholder="exemple@domaine.com" required>
+                    </div>
+                    <div class="form-group">
+                        <label for="userPassword">Mot de passe :</label>
+                        <input type="password" class="form-control" id="userPassword" name="userPassword" placeholder="exemple" required>
+                    </div>
+                    <button type="submit" class="btn btn-default">Se connecter</button>
+                    <a href="register" class="pull-right">S'enregister</a>
+                </form>
+            </div>
+            <div class="col-lg-3 col-sm-pull-5  col-sm-6">
+                <img class="img-responsive" src="img/connection.png" alt="">
+            </div>
+        </div>
+
+    </div>
+    <!-- /.container -->
+
+</div>
+<!-- /.content-section-b -->
+
+<!-- jQuery -->
+<script src="js/jquery.js"></script>
+
+<!-- Bootstrap Core JavaScript -->
+<script src="js/bootstrap.min.js"></script>
+
 </body>
+
 </html>
